+++ app/controllers/backend/content/rows/columns_controller.rb
@@ -1,6 +1,6 @@
 class Backend::Content::Rows::ColumnsController < BackendController
   before_action :find_row
-  before_action :find_column, only: [:edit, :update, :destroy]
+  before_action :find_column, only: [:edit, :update, :move_up, :move_down, :destroy]
   layout 'backend/lightbox'
 
   def new
@@ -29,6 +29,16 @@ class Backend::Content::Rows::ColumnsController < BackendController
     end
   end
 
+  def move_up
+    @column.move_higher
+    redirect_to cancel_url
+  end
+
+  def move_down
+    @column.move_lower
+    redirect_to cancel_url
+  end
+
   def destroy
     @column.destroy
 
@@ -51,15 +61,7 @@ class Backend::Content::Rows::ColumnsController < BackendController
   end
 
   def redirect_to_edit(column)
-    if column.content.is_a?(ContentText)
-      redirect_to edit_backend_content_text_path(column.content)
-
-    elsif column.content.is_a?(ContentImage)
-      redirect_to edit_backend_content_image_path(column.content)
-
-    else
-      raise "No such content type #{column.content.class}"
-    end
+    redirect_to send("edit_backend_content_#{column.content.decorate.content_type}_path", column.content)
   end
 
   def cancel_url