+++ app/decorators/page_decorator.rb
@@ -9,11 +9,23 @@ class PageDecorator < Draper::Decorator
     h.options_for_select(list, selected: parent_id, disabled: disabled)
   end
 
-  # TODO calculate the actual path if no route is set in the page
-  # TODO take into account that some sites don't use a language prefix!
-  # this might mean that we need to configure this in the udongo config.
-  # eg prefix_routes_with_locale = false (default = true)
-  def path(options = {})
-    route.present? ? h.send(route, options) : '/'
+  def path(locale: I18n.locale, options: {})
+    return h.send(route, options) if route.present?
+
+    slugs = []
+
+    parents.each do |p|
+      if p.route.present?
+        slugs << h.send(p.route, options)
+        return slugs.reverse.join('/')
+      else
+        slugs << p.seo(locale).slug
+      end
+    end
+
+    str = ''
+    str << "/#{locale}" if Udongo.config.prefix_routes_with_locale?
+    str << "/#{slugs.reverse.join('/')}"
+    str
   end
 end