+++ app/models/concerns/translatable.rb
@@ -11,8 +11,12 @@ module Concerns
     end
 
     def translation(locale = I18n.locale)
-      @collections = {} unless @collections
-      @collections[locale.to_sym] ||= Collection.new(self, self.class.translation_config, locale)
+      @translation_collections = {} unless @translation_collections
+      @translation_collections[locale.to_sym] ||= Collection.new(self, self.class.translation_config, locale)
+    end
+
+    def translatable?
+      true
     end
 
     module ClassMethods