+++ lib/udongo/breadcrumb.rb
@@ -1,4 +1,3 @@
-
 module Udongo
   class Breadcrumb
     def initialize