+++ spec/factories/navigations.rb
@@ -1,6 +1,6 @@
 FactoryGirl.define do
   factory :navigation do
-    sequence(:name) { |n| "foo-#{n}" }
+    sequence(:identifier) { |n| "foo-#{n}" }
     description 'foo'
   end
 end